Moved <script> to container file, created proper class file.
The script tag can't be used in the same way as the stylesheet is used.  To that end, I've moved it into its own class and put it in the container file, with appropriate modifications ot the other files (drawing / element factory).  I've tested and it seems to work fine.

I also added an appropriate test file to show the expected output.

I don't suppose anyone will actually want to put their content directly into the SVG file so I haven't added that functionality.